requirements.txt
paramiko==4.0.0
PyYAML==6.0.3
click==8.3.0